+++ weather-tvOS/Model/WeatherData.swift
